test(jianying): cover effect catalog parsing
